[C] Have add and remove destination return the correlation id from the client conductor for compatibility with C++ API.

+++ aeron-client/src/main/c/aeron_client_conductor.h
@@ -322,7 +322,7 @@ int aeron_client_conductor_linger_image(aeron_client_conductor_t *conductor, aer
 int aeron_client_conductor_offer_remove_command(
     aeron_client_conductor_t *conductor, int64_t registration_id, int32_t command_type);
 
-int aeron_client_conductor_offer_destination_command(
+int64_t aeron_client_conductor_offer_destination_command(
     aeron_client_conductor_t *conductor, int64_t registration_id, int32_t command_type, const char *uri);
 
 inline int aeron_counter_heartbeat_timestamp_find_counter_id_by_registration_id(
